refactor: create utils
